WIP: Fixing tests
testExtendedErrorMessage -- Depening on the browser, 'ID' or 'id' will be used. Thus, removed id from checked string
testAddTestCookie -- Depending on the browser, the Path was set to "" or "/". When testing Cookies, I do not think
we care about the Path assigned by the browser, so I removed the check for the Path